feat: Add input for release note and ignore dependecies
'changes' input will allow to specify the breif release note that is
added to jmeter-plugins.

'ingore-dependecies' will allow to avoid publishing to jmeter-plugins
vertain artifacts that are attached into the github release but aren't
required to be added in jmeter-plugins release. There might be some
interesting use cases for this particular input. The initial development
purpose was to avoid adding the bzm-repositories library currently
attached to some of the ACR github releases, as a plugin library in the
jmeter-plugins release. The bzm-repositories it's a plugin by itself
that does not have at the moment a place for the jar file to be release
other than the ACR github release. Mainly becouse it's a propietary and
sensitive development that can't reach public access.