fix: Force default value for jmeter-plugins url input
Even though the default value is set in the action.yaml definition when
running locally it does not bring back the default value for the desired
input.

Notice in the action.yaml that the input is not mandatory either.

Github does not explicity tell how this is supoused to work. Locally
default non required values aren't working.