feat: Add checkout release branch
Build the actual name for the release branch.

Along with that. We update the release branch with the latest changes
possible from the upstream. For now it's fixed to grab master since
undera/ uses master as his main branch.

Additionally, expose the extraction of owner and repo from a URL due to
usage outside of this file. It would be possible to remove this function
implementation into a utils.ts class in the future.