feat: Add open pull request
Use github service through github rest api in order to open a Pull
Request inside of the upstream repository from the forked release
branch.
Title of the PR is set based on the current plugin name and release
Body of the current PR is based on the github release notes for the
released version.

NOTE: The PR is set as draft as a contingency plan in case soemthing
went wrong and it wasn't captured in the whole cycle.