fix: Consume github service as js module
Since we are using ncc to bundle the code. There is an issue when
enabling `"allowImportingTsExtensions": true` in typescript config which
won't recognize `noEmit:true` set and it will fail the boundling saying
something in the lines of missing noEmit key.

Therefore, there is a reason to back up on the allowImportingTsExtension
and keep using the esm standard of import which will be by adding the js
extension even though we are using ts.