fix: Improve filtering of ignored-dependecies
There was an issue related to ignoring dependencies causing that none of
the libs were processed when ingoredDependecies was empty. Filtering
adjusted in order to satify that it won't match every asset name when
empty.
Every string startsWith empty string, therefore we were excluding all
assets when ingore dependencies empty